+++ tests/python/openassetio/hostAPI/test_session.py
@@ -270,6 +270,60 @@ class Test_Session_createContext:
             a_host_session, parentState=state_a)
 
 
+class Test_Session_freezeContext:
+
+    def test_when_called_with_no_manager_then_RuntimeError_is_raised(self, a_session):
+
+        with pytest.raises(RuntimeError):
+            a_session.freezeContext(Context())
+
+    def test_when_called_then_the_managers_frozen_token_is_returned(
+             self, a_session, mock_manager_interface):
+
+        expected_token = "a_frozen_token"
+        mock_manager_interface.mock.freezeState.return_value = expected_token
+
+        a_session.useManager("com.manager")
+
+        initial_state = "initial_state"
+        a_context = Context()
+        a_context.managerInterfaceState = initial_state
+
+        actual_token = a_session.freezeContext(a_context)
+
+        assert actual_token == expected_token
+
+        # pylint: disable=protected-access
+        a_host_session = a_session.currentManager()._Manager__hostSession
+        mock_manager_interface.mock.freezeState.assert_called_once_with(
+            initial_state, a_host_session)
+
+
+class Test_Session_thawContext:
+
+    def test_when_called_with_no_manager_then_RuntimeError_is_raised(self, a_session):
+
+        with pytest.raises(RuntimeError):
+            a_session.thawContext(Context())
+
+    def test_when_called_then_the_managers_thawed_state_is_set_in_the_context(
+             self, a_session, mock_manager_interface):
+
+        expected_state = "thawed_state"
+        mock_manager_interface.mock.thawState.return_value = expected_state
+
+        a_session.useManager("com.manager")
+
+        a_token = "frozen_token"
+        a_context = a_session.thawContext(a_token)
+
+        assert a_context.managerInterfaceState is expected_state
+
+        # pylint: disable=protected-access
+        a_host_session = a_session.currentManager()._Manager__hostSession
+        mock_manager_interface.mock.thawState.assert_called_once_with(a_token, a_host_session)
+
+
 class Test_Session_getSettings:
 
     def test_when_called_with_no_manager_then_identifier_key_is_empty(self, a_session):